gcodepainter/requirements.txt
Cython
kivy
PyDispatcher
PySerial
